requirements.txt
numpy>=1.24.0
PyYAML>=6.0.1
tqdm>=4.65.0
requests>=2.31.0
pandas>=2.0.0

torch>=2.0.0
torchvision>=0.15.0

transformers>=4.30.0
openai==0.28.0

osmnx>=1.6.0
geopandas>=0.13.0
shapely>=2.0.0

selenium>=4.10.0
webdriver-manager>=4.0.0
geopy>=2.4.0